Honor coding guidelines (naming, white-spaces) & do an optimization.
About the previous commit : I thought PR could not be overwritten
because it is used into the "interrupted" function, such as R0.
However, PR is always restored (from the stack) into epilogues and
is only used at this moment, so it can be used safely (exception
can not occur into epilogues).

About this commit : Adjust some lines in order to keep the code
consistent and to ensure prettiness. Also, do the following
optimization :

    add offsetof(MonoContext, pc), sh4_r0
    mov.l @sh4_r0, sh4_r0

is converted to :

    mov.l @(offsetof(MonoContext, pc), sh4_r0), sh4_r0