Add initial mono_arch_lowering_pass() + support for sh4_icompare_imm_R0.
The function mono_arch_lowering_pass() decomposes some generic opcodes
to architecture-specific ones. It is really helpful because it gives to
Mono a better estimation of used resources (registers and length).

For instance, the opcode "icompare_imm" is not easily translatable
from MonoInst to SH4 code because it only works with R0, moreover
the global register allocator and the "call" handler have already
done some register assignment. This lowering pass allows to do
something like this :

    if ins->sreg1 is not already assigned to a physical register and
       ins->inst_imm is in the range of "cmpeq_imm_R0" then
        replace OP_ICOMPARE_IMM with OP_SH4_ICOMPARE_IMM_R0
    else
        replace OP_ICOMPARE_IMM with OP_ICOMPARE