Finalize and fix the support to exceptions for SH4.
From an IR point of view, the opcodes "cond_exec_*" for SH4 are
handled as other conditional branches, and because these are flattened
to "sh4_bt/f" opcodes, we have to remember the "exceptional-ness"
(inst->backend.data == -1) to set correctly the patch_info into
output_basic_block().

Note1: Currently, we do not support the reuse of a throw sequence for
the same exception class.

Note2: As a side effect of this patch, the opcodes "bgt_un" and
"ble_un" are now supported.