SH4: fix the register allocation for "far" (store|load)_membase opcodes.
Previously a "load_membase" with a big offset was lowered to something
like:

    iconst Roffset <- big_offset
    int_add Rbase <- Rbase Roffset
    sh4_loadi4 Rvalue <- Rbase

As you can see the base register was overwritten. This is a problem
when this base register is the frame pointer (for instance). Now, the
lowering is something like:

    iconst Roffset <- big_offset
    int_add Roffset <- Rbase Roffset
    sh4_loadi4 Rvalue <- Roffset

It works fine because Roffset is a newly created virtual register.

The function get_vcall_slot_addr() was adjusted according to this new
register allocation.