SH4: the constant-pool is now emitted after the native instructions.
This patch prepares the arrival of the "full perf" constant-pool
manager. Previously, the constant-pool was emitted in place, for
instance the opcode "call" produced the following code sequence:

     mov.l   @(PC+6), r3
     bra     PC+8
     nop
     .word1
     .word2
     jsr     @r3
     nop

Now, it produces:

     mov.l   @(PC+6), r3
     jsr     @r3
     nop
     bra     PC+8
     nop
     .word1
     .word2

As you can guess, this patch allows to put several constants into
one pool (currently we are still doing 1:1).