MIPS: [HACK] Set process affinity to CPU 0 to avoid multi-core MIPS crashes
The current MIPS port suffers from stability issues when running on a
true multiprocessor platform; it appears that the code generator or
runtime is violating the MIPS memory model by not using enough/correct
fencing instructions.

The current workaround is to pin the Mono process to execute on one
processor only; this results in a stable (if not optimal) runtime.