[HACK] Xamarin/46567, Async-local Culture/UICulture
This is a minimal, proof-of-concept patch which reuses the async-local
support from MS' implementation to implement our use-case.  I do not
consider the solution "clean" by any means, but we needed a quick fix on
top of mono-4.6.1.3, and this seems to fit the bill.  Uploading it cases
it helps somebody else.