SH4: Allow generation of NATIVE_TO_MANAGED wrappers (mono-sh4#5)
     https://code.google.com/archive/p/mono-sh4/issues/5

This commit "fixes" the issue in the sense that no assertion is
triggered.

Many wrappers seem to work, but some tests such as pinvoke2.exe,
winx64structs.exe and pinvoke3.exe still crash the runtime, because
the current P/Invoke code does not implement the following part of the
SH-4 ABI:

> The parameter is broken into a series of 4-byte elements. If there
> are enough registers available from the list R4, R5, R6, R7 to
> contain all of the 4-byte elements of the parameter, the 4-byte
> elements are assigned to the available registers in order.

causing one of the side to see (and choke on) garbage.