SH4: Use mini_jit_info_table_find when unwinding native frames
A crucial difference between 'mono_jit_info_table_find,' which was
used before this patch, and the mini_* variant is that the latter will
find JIT metadata for code whose PC is not in the current domain.

Without this, the unwinding of a cross-domain stack can fail (i.e.,
return FALSE), causing 'is_running_protected_wrapper' to potentially
wrongly return FALSE--with the consequence that protected wrappers can
be unwound.  Oops!

This fixes 'appdomain-thread-abort.exe' on the SH-4 backend.