SH4: Do not use SP as scratch register when popping LMFs
This commit fixes an intermittent crash caused by the JIT using SP as
a scratch register, for a window of three instructions, with proper
saving and restoring afterwards--an "optimisation" which is invalid in
a preemptively multithreaded GCed runtime, as this can cause the GC to
see bogus stack ranges.

Let's just use r4 (whose value should not matter here) instead.