WIP [runtime] Forcibly clean LMF on Threading.Monitor.Monitor_wait entry
'Monitor.Wait' is implemented by a native 'Monitor_wait' internal
call, and that "icall" can sometimes be very long-lived.

As with other transitions from managed to native, this results in the
allocation of a "last managed frame" (LMF) data structure on the
thread's stack.  The lifetime of that record naturally matches that of
the call.

On ARM, the LMF includes space for saving floating-point registers,
but the space is only used during trampolines, not for "normal"
transitions.  This means that on entry, that 'fregs' field is left
uninitialized, and just overlaid on the stack--"capturing" whichever
words happen to stored at that location.

Mono's current garbage collector scans the stack conservatively, and
is unaware of the 'double[]' nature of the 'fregs' field.  It scans
word by word, and interprets all values as "potential pointers."

On an (a priori) unrelated note, Mono's current 'ThreadPoolWorkQueue',
inherited from Microsoft's .NET, keeps some "work items" in a lock-free
singly-linked list of arrays, made of 'QueueSegment' instances.  It
holds advancing "head" and "tail" pointers into that list, and expects
the GC to release all segments preceding the current tail.

The points above, combined, can result in a memory leak.

In an observed case, a well-defined scenario caused the 'Monitor_wait'
LMF to systematically capture a 'QueueSegment' pointer--effectively
pinning the list and inducing unbounded memory growth for the duration
of the call.

Early signs looked like this in the profiler log:

    408         17       24 System.Threading.ThreadPoolWorkQueue.QueueSegment (bytes: +384, count: +16)

(The 'IThreadPoolWorkItem[]' growth was more visible/alarming, but a
direct consequence of the above.)

In a test run, 0x76b33c10 was the first retained 'QueueSegment' pointer;
that word could found in the LMF's floating point registers field:

    (gdb) x/32wx $lmf->fregs
    0x5bbfe6a8:     0x5bbfe828      0x727431a0      0x76becd00      0x76b75938
    [...]
    0x5bbfe6e8:     0x76b33c10      0x00000002      0x00d56200      0x76c4cee0
                    ^^^^^^^^^^

The pointer values varied from run to run, but the phenomenon was
quite reproducible.  Here is a representative backtrace:

    "[Redacted]" tid=0x5aaff440 this=0x5bca42b0 , thread handle : 0x5b2fa148, state : waiting
      at <unknown> <0xffffffff>
      at (wrapper managed-to-native) System.Threading.Monitor.Monitor_wait (object,int) <0x0003b>
      at System.Threading.Monitor.ObjWait (bool,int,object) [0x0002f] in <b23d5b7bf8d949308b1e73dcadb8de10>:0
      at System.Threading.Monitor.Wait (object,int,bool) [0x0000e] in <b23d5b7bf8d949308b1e73dcadb8de10>:0
      at System.Threading.Monitor.Wait (object,int) [0x00000] in <b23d5b7bf8d949308b1e73dcadb8de10>:0
      at System.Threading.SemaphoreSlim.WaitUntilCountOrTimeout (int,uint,System.Threading.CancellationToken) [0x0001d] in <b23d5b7bf8d949308b1e73dcadb8de10>:0
      at System.Threading.SemaphoreSlim.Wait (int,System.Threading.CancellationToken) [0x000d9] in <b23d5b7bf8d949308b1e73dcadb8de10>:0
      at System.Collections.Concurrent.BlockingCollection`1<T_REF>.TryTakeWithNoTimeValidation (T_REF&,int,System.Threading.CancellationToken,System.Threading.CancellationTokenSource) [0x00067] in <5b53ebaedc3e458db80f9e7e4412b1f1>:0
      at System.Collections.Concurrent.BlockingCollection`1/<GetConsumingEnumerable>d__68<T_REF>.MoveNext () [0x0004d] in <5b53ebaedc3e458db80f9e7e4412b1f1>:0
      [...]

This "WIP" patch works around the problem in the ARM backend by
special-casing the 'Monitor_wait' method/icall, and systematically
clearing its LMF on entry.

This is an ugly hack & workaround, not a "correct" solution: the
latter would require either precise stack-scanning by the GC (ideal),
or fully initialized LMFs during managed/native transitions (costly
and not as robust).

We do not intend to polish the patch as this point, as Mono 5.18
"realigned the planets" (or, rather, stack frame offsets)--which
incidentally got rid of the issue.  The above was just "bad luck," and
that bad luck is gone.

But it could come back at any point, for anybody--until the Mono GC is
updated/switched to precisely scan the stack.  The analysis above, if
not the patch, may help the poor souls!